requirements.txt
Flask
requests
beautifulsoup4
nltk
pytesseract
Pillow
pdfplumber
python-docx
imapclient
langdetect
